Refine PMTiles archive state handling
