+++ pmtiles/src/commonTest/kotlin/org/maplibre/spatialk/pmtiles/internal/DirectoryDecodingTest.kt
@@ -66,6 +66,7 @@ class DirectoryDecodingTest {
 
     @Test
     fun rejectsMalformedDirectories() {
+        assertDirectoryFails(PmTilesErrorCode.InvalidDirectory, bytes(0uL))
         assertDirectoryFails(PmTilesErrorCode.InvalidDirectory, bytes(1uL, 0uL, 1uL, 0uL, 1uL))
         assertDirectoryFails(PmTilesErrorCode.InvalidDirectory, bytes(1uL, 0uL, 1uL, 1uL, 0uL))
         assertDirectoryFails(PmTilesErrorCode.InvalidDirectory, bytes(2uL, 5uL, 0uL))
