+++ pmtiles/src/commonTest/kotlin/org/maplibre/spatialk/pmtiles/internal/LimitsTest.kt
@@ -3,6 +3,7 @@ package org.maplibre.spatialk.pmtiles.internal
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
+import org.maplibre.spatialk.pmtiles.ArchiveLimits
 import org.maplibre.spatialk.pmtiles.ArchiveSection
 import org.maplibre.spatialk.pmtiles.ByteRange
 import org.maplibre.spatialk.pmtiles.PmTilesErrorCode
@@ -53,5 +54,41 @@ class LimitsTest {
             }
 
         assertEquals(PmTilesErrorCode.LimitExceeded, error.code)
+    }
+
+    @Test
+    fun rejectsInvalidArchiveLimitConfiguration() {
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits(maxInitialReadBytes = 16 * 1024 - 1)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits(maxMetadataBytes = -1)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits.Default.copy(maxDirectoryDecompressedBytes = 1024)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits(maxDirectoryDecompressedBytes = 1024, maxDirectoryEntries = 256)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits(maxDirectoryEntries = -1)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits(maxDirectoryEntries = Int.MAX_VALUE)
+        }
+        assertFailsWith<IllegalArgumentException> {
+            ArchiveLimits(maxVarintBytes = 0)
+        }
+    }
+
+    @Test
+    fun acceptsDirectoryEntryLimitThatFitsDecompressedDirectoryBudget() {
+        val limits =
+            ArchiveLimits(
+                maxDirectoryDecompressedBytes = 1024,
+                maxDirectoryEntries = 255,
+            )
+
+        assertEquals(255, limits.maxDirectoryEntries)
     }
 }
